refactor: 웨이팅 로직 분리
